requirements.txt
asgiref==3.5.0
Django==4.0.4
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
pytesseract==0.3.9
sqlparse==0.4.2
